refactor(ports): move Sleipnir to standalone src/sleipnir module
Move all Sleipnir code out of niuu and into its own top-level module
under src/sleipnir/, following the same hexagonal structure:

  src/sleipnir/
  ├── domain/events.py        SleipnirEvent + event type registry
  ├── ports/events.py         Publisher, Subscriber, Subscription ABCs
  └── adapters/
      ├── in_process.py       asyncio in-memory transport
      └── serialization.py    msgpack + JSON wire formats

Tests move from tests/test_niuu/ to tests/test_sleipnir/.
Import paths updated from niuu.* to sleipnir.*.
pyproject.toml: add src/sleipnir to wheel packages and coverage source.
