--- src/niuu/adapters/sleipnir/__init__.py
@@ -1,1 +1,0 @@
-"""Sleipnir event bus adapters."""
