+++ src/sleipnir/__init__.py
@@ -1,0 +1,6 @@
+"""Sleipnir — event bus for the Niuu platform.
+
+Sleipnir provides a structured publish/subscribe event bus used across
+Ravn, Tyr, Volundr, and Bifrost. The contract (SleipnirEvent, port
+interfaces) is transport-agnostic; adapters provide the actual delivery.
+"""
