+++ src/sleipnir/adapters/__init__.py
@@ -1,0 +1,1 @@
+"""Sleipnir transport adapters."""
