+++ src/sleipnir/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Sleipnir domain models."""
